Move component files to dirs
Closes #28
