+++ src/components/Legal/index.tsx
