+++ src/components/Tag/index.tsx
@@ -1,7 +1,7 @@
 import { graphql, Link } from "gatsby";
 import React from "react";
 import Helmet from "react-helmet";
-import { Layout } from "../components/Layout";
+import { Layout } from "../Layout";
 
 export interface ITagProps {
   totalCount: number;
